services/user-service/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Gor903", email = "gor.beglaryan.rw@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
packages = [
    { include = "user_service", from = "src" }
]
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
